Modified Makefiles and Menu and CPP conversion
