+++ Lab1/problem2a_immediate.cpp
@@ -1,4 +1,4 @@
-#include <stdio.h>
+#include <iostream>
 #include <mpi.h>
 
 int main(int argc, char** argv) {
@@ -7,7 +7,6 @@ int main(int argc, char** argv) {
     int nproc = 1;
     int namelength;
     char processor_name[MPI_MAX_PROCESSOR_NAME];
-    int name_len;
 
     MPI_Status status;
     MPI_Request request;
@@ -15,26 +14,26 @@ int main(int argc, char** argv) {
     MPI_Init (&argc, &argv); // Initialize MPI
     MPI_Comm_size (MPI_COMM_WORLD, &nproc); // Get the number of processors
     MPI_Comm_rank (MPI_COMM_WORLD, &rank); // Get my rank
-    MPI_Get_processor_name(processor_name, &name_len);// Get the name of the processor and length
+    MPI_Get_processor_name(processor_name, &namelength);// Get the name of the processor and length
 
     if(rank==0)
     {
         char message_buffer[MPI_MAX_PROCESSOR_NAME];
         
-        printf("Number of processors is %d\n", nproc);
-        printf("Node %d is processor %s\n",rank, processor_name);
+        std::cout << "Number of processors is " << nproc << std::endl;
+        std::cout << "Node " << rank <<" is processor " << processor_name << std::endl;
 
         for(int irank = 1; irank < nproc; irank++)
         {
             //Receive name from processor with rank = irank - store in "message_queue"
             MPI_Irecv(&message_buffer, MPI_MAX_PROCESSOR_NAME, MPI_CHAR, irank, 99, MPI_COMM_WORLD, &request);
             MPI_Wait (&request, &status);
-            printf("Node %d receiving message \"%s\" from Node %d\n", rank, message_buffer, irank);
+            std::cout << "Node " << rank << " receiving message:"  << message_buffer  << " from Node" << irank << std::endl;
         }
     }
     else
     {
-        printf("Node %d sending \"%s\" to Node %d\n", rank, processor_name, 0);
+        std::cout << "Node " << rank << " sending:"  << processor_name  << " to Node" << 0 << std::endl;
         MPI_Isend(&processor_name, MPI_MAX_PROCESSOR_NAME, MPI_CHAR, 0, 99, MPI_COMM_WORLD, &request); // Send processor_n ame to processor 0.
         MPI_Wait(&request, MPI_STATUS_IGNORE) ;// Wait for MPI_Isend to complete before progressing further
     }
